@@ -10,17 +10,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.5",
+    "@angular/animations": "^18.1.3",
     "@angular/cdk": "^17.2.2",
-    "@angular/common": "^17.0.5",
-    "@angular/compiler": "^17.0.5",
-    "@angular/core": "^17.0.5",
+    "@angular/common": "^18.1.3",
+    "@angular/compiler": "^18.1.3",
+    "@angular/core": "^18.1.3",
     "@angular/fire": "^17.0.1",
-    "@angular/forms": "^17.0.5",
+    "@angular/forms": "^18.1.3",
     "@angular/material": "^17.2.2",
-    "@angular/platform-browser": "^17.0.5",
-    "@angular/platform-browser-dynamic": "^17.0.5",
-    "@angular/router": "^17.0.5",
+    "@angular/platform-browser": "^18.1.3",
+    "@angular/platform-browser-dynamic": "^18.1.3",
+    "@angular/router": "^18.1.3",
     "@ng-icons/core": "^26.1.0",
     "@ng-icons/heroicons": "^26.1.0",
     "ngx-infinite-scroll": "^17.0.0",
@@ -29,9 +29,9 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.5",
-    "@angular/cli": "~17.0.5",
-    "@angular/compiler-cli": "^17.0.5",
+    "@angular-devkit/build-angular": "^18.1.3",
+    "@angular/cli": "~18.1.3",
+    "@angular/compiler-cli": "^18.1.3",
     "@types/jasmine": "~4.3.0",
     "autoprefixer": "^10.4.16",
     "jasmine-core": "~4.6.0",
@@ -42,6 +42,6 @@
     "karma-jasmine-html-reporter": "~2.1.0",
     "postcss": "^8.4.30",
     "tailwindcss": "^3.3.3",
-    "typescript": "~5.2.2"
+    "typescript": "~5.5.4"
   }
 }
