@@ -1,7 +1,7 @@
 // Angular core libraries
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
 
 // Application components
@@ -16,19 +16,12 @@ import { PlaylistsModule } from './my-playlists/playlists.module';
 import { SharedModule } from './shared/shared.module';
 import { CoreModule } from './core.module';
 
-@NgModule({
-  declarations: [AppComponent, HeaderComponent],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    HttpClientModule,
-    FirebaseModule,
-    PlaylistsModule,
-    SharedModule,
-    CoreModule,
-    FooterComponent,
-  ],
-  providers: [provideAnimationsAsync('noop')],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [AppComponent, HeaderComponent],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        FirebaseModule,
+        PlaylistsModule,
+        SharedModule,
+        CoreModule,
+        FooterComponent], providers: [provideAnimationsAsync('noop'), provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule {}
